+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master
 
+* Support `params` special parameter
+
 ## v2.2.4
 
 * Fix rails engine loading if sprockets is not in Gemfile. Fixes [#294](https://github.com/railsware/js-routes/issues/294)